+++ pkg/providers/property/data_akamai_cp_code.go
@@ -3,6 +3,7 @@ package property
 import (
 	"context"
 	"fmt"
+	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -62,8 +63,8 @@ func dataSourceCPCode() *schema.Resource {
 
 func dataCPCodeRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	meta := akamai.Meta(m)
+	client := inst.Client(meta)
 	log := meta.Log("PAPI", "dataCPCodeRead")
-
 	log.Debug("Read CP Code")
 
 	var name, groupID, contractID string
@@ -97,7 +98,7 @@ func dataCPCodeRead(ctx context.Context, d *schema.ResourceData, m interface{}) 
 		return diag.Errorf("%v: %s", tf.ErrValueSet, err.Error())
 	}
 
-	cpCode, err := findCPCode(ctx, name, contractID, groupID, meta)
+	cpCode, err := findCPCode(ctx, client, name, contractID, groupID)
 	if err != nil {
 		return diag.FromErr(fmt.Errorf("could not load CP codes: %w", err))
 	}
@@ -109,15 +110,14 @@ func dataCPCodeRead(ctx context.Context, d *schema.ResourceData, m interface{}) 
 	if err := d.Set("product_ids", cpCode.ProductIDs); err != nil {
 		return diag.Errorf("%v: %s", tf.ErrValueSet, err.Error())
 	}
-	d.SetId(cpCode.ID)
 
+	d.SetId(strings.TrimPrefix(cpCode.ID, cpCodePrefix))
 	log.Debugf("Read CP Code: %+v", cpCode)
 	return nil
 }
 
 // findCPCode searches all CP codes for a match against given nameOrID
-func findCPCode(ctx context.Context, nameOrID, contractID, groupID string, meta akamai.OperationMeta) (*papi.CPCode, error) {
-	client := inst.Client(meta)
+func findCPCode(ctx context.Context, client papi.PAPI, nameOrID, contractID, groupID string) (*papi.CPCode, error) {
 	r, err := client.GetCPCodes(ctx, papi.GetCPCodesRequest{
 		ContractID: contractID,
 		GroupID:    groupID,
@@ -128,10 +128,14 @@ func findCPCode(ctx context.Context, nameOrID, contractID, groupID string, meta 
 	}
 
 	for _, cpc := range r.CPCodes.Items {
-		if cpc.ID == nameOrID || cpc.ID == "cpc_"+nameOrID || cpc.Name == nameOrID {
+		if cpCodeNameOrIDMatches(cpc, nameOrID) {
 			return &cpc, nil
 		}
 	}
 
 	return nil, fmt.Errorf("%w: CP code: %s", ErrCpCodeNotFound, nameOrID)
+}
+
+func cpCodeNameOrIDMatches(cpCode papi.CPCode, s string) bool {
+	return cpCode.ID == s || cpCode.ID == "cpc_"+s || cpCode.Name == s
 }