+++ pkg/providers/property/resource_akamai_cp_code.go
@@ -92,8 +92,11 @@ var (
 	cpCodeResourceUpdateTimeout = time.Minute * 30
 )
 
+const cpCodePrefix = "cpc_"
+
 func resourceCPCodeCreate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	meta := akamai.Meta(m)
+	client := inst.Client(meta)
 	logger := meta.Log("PAPI", "resourceCPCodeCreate")
 	logger.Debugf("Creating CP Code")
 
@@ -103,35 +106,33 @@ func resourceCPCodeCreate(ctx context.Context, d *schema.ResourceData, m interfa
 	}
 
 	// Schema guarantees product_id/product are strings and one or the other is set
-	productID := d.Get("product_id").(string)
-	if productID == "" {
+	var productID string
+	if got, ok := d.GetOk("product_id"); ok {
+		productID = got.(string)
+	} else {
 		productID = d.Get("product").(string)
-		if productID == "" {
-			return diag.Errorf("one of product,product_id must be specified")
-		}
 	}
 	productID = tools.AddPrefix(productID, "prd_")
 
 	contractID, groupID := getContractIDAndGroupID(d)
 
+	var cpCodeID string
 	// Because CPCodes can't be deleted, we re-use an existing CPCode if it's there
-	cpCode, err := findCPCode(ctx, name, contractID, groupID, meta)
+	cpCode, err := findCPCode(ctx, client, name, contractID, groupID)
 	if err != nil && !errors.Is(err, ErrCpCodeNotFound) {
 		return diag.Errorf("%s: %s", ErrLookingUpCPCode, err)
 	}
 
-	if cpCode == nil {
-		cpcID, err := createCPCode(ctx, name, productID, contractID, groupID, meta)
+	if errors.Is(err, ErrCpCodeNotFound) {
+		cpCodeID, err = createCPCode(ctx, client, name, productID, contractID, groupID)
 		if err != nil {
 			return diag.FromErr(err)
 		}
-
-		d.SetId(cpcID)
 	} else {
-		d.SetId(cpCode.ID)
+		cpCodeID = cpCode.ID
 	}
 
-	logger.Debugf("Resulting CP Code: %#v", cpCode)
+	d.SetId(strings.TrimPrefix(cpCodeID, cpCodePrefix))
 	return resourceCPCodeRead(ctx, d, m)
 }
 
@@ -180,7 +181,7 @@ func resourceCPCodeRead(ctx context.Context, d *schema.ResourceData, m interface
 	if err := d.Set("product_id", cpCode.ProductIDs[0]); err != nil {
 		return diag.Errorf("%s: %s", tf.ErrValueSet, err.Error())
 	}
-	d.SetId(cpCode.ID)
+
 	logger.Debugf("Read CP Code: %+v", cpCode)
 	return nil
 }
@@ -198,7 +199,8 @@ func resourceCPCodeUpdate(ctx context.Context, d *schema.ResourceData, m interfa
 
 	contractID, groupID := getContractIDAndGroupID(d)
 
-	cpCodeID, err := strconv.Atoi(strings.TrimPrefix(d.Id(), "cpc_"))
+	// trimCPCodeID is needed here for backwards compatibility
+	cpCodeID, err := strconv.Atoi(strings.TrimPrefix(d.Id(), cpCodePrefix))
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -250,7 +252,7 @@ func resourceCPCodeImport(ctx context.Context, d *schema.ResourceData, m interfa
 	if parts[0] == "" {
 		return nil, errors.New("CP Code is a mandatory parameter")
 	}
-	cpCodeID := tools.AddPrefix(parts[0], "cpc_")
+	cpCodeID := parts[0]
 	contractID := tools.AddPrefix(parts[1], "ctr_")
 	groupID := tools.AddPrefix(parts[2], "grp_")
 
@@ -283,14 +285,14 @@ func resourceCPCodeImport(ctx context.Context, d *schema.ResourceData, m interfa
 	if err := d.Set("product", cpCode.ProductIDs[0]); err != nil {
 		return nil, fmt.Errorf("%w: %s", tf.ErrValueSet, err.Error())
 	}
-	d.SetId(cpCode.ID)
+
+	d.SetId(strings.TrimPrefix(cpCode.ID, cpCodePrefix))
 	logger.Debugf("Import CP Code: %+v", cpCode)
 	return []*schema.ResourceData{d}, nil
 }
 
 // createCPCode attempts to create a CP Code and returns the CP Code ID
-func createCPCode(ctx context.Context, name, productID, contractID, groupID string, meta akamai.OperationMeta) (string, error) {
-	client := inst.Client(meta)
+func createCPCode(ctx context.Context, client papi.PAPI, name, productID, contractID, groupID string) (string, error) {
 	r, err := client.CreateCPCode(ctx, papi.CreateCPCodeRequest{
 		ContractID: contractID,
 		GroupID:    groupID,