+++ pkg/providers/property/testdata/TestDSCPCode/match_by_full_id.tf
@@ -3,7 +3,7 @@ provider "akamai" {
 }
 
 data "akamai_cp_code" "test" {
-  name     = "cpc_test2"
+  name     = "cpc_234"
   contract = "ctr_test"
   group    = "grp_test"
 }